# docs/source/_openapi/device_openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/device/health/readiness":{"get":{"tags":["health"],"summary":"Readiness","description":"Readiness health endpoint.\n\nInspects SQLAlchemy engine and checks if the device table exists.","operationId":"health_readiness","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/device/":{"get":{"tags":["devices"],"summary":"Get Devices","description":"Retrieve the list of registered devices.\n\nReturns\n-------\n    List[Device]: The list of registered devices.","operationId":"get_devices","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DeviceOut"},"type":"array","title":"Response Get Devices"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/device/{device_id}":{"get":{"tags":["devices"],"summary":"Get Device","description":"Retrieve a specific device.\n\nArgs\n-------\n    device_id (str): The ID of the device.\n\nReturns\n-------\n    dict: The response containing the information about the device","operationId":"get_device","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"device_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Device Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["devices"],"summary":"Delete Device","description":"Delete a device.\n\nArgs\n-------\n    device_id (str): The ID of the device.\n\nReturns\n-------\n    dict: The response indicating the success or failure of the deletion.","operationId":"delete_device","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"device_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Device Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/device/createdevice":{"post":{"tags":["devices"],"summary":"Create Device","description":"Create device database entry.\n\nOnly devices which are created by a user via this endpoint and have the device_id and device_token\nthat result from this call configured correctly can later connect and register.\n\nParameters\n----------\nrequest:\n    DeviceCreationRequest to create device in Scanhub.\n    Further details of the device are later provided by the device itself.\n\nReturn\n------\ndevice_token:\n    the token of the device which the user should copy manually to the configuration file of the device.\n\ndevice_id:\n    the id of the device, should be copied to the devices config file together with the device_token.","operationId":"create_device","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceCreationRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/device/parameter/{device_id}":{"put":{"tags":["devices"],"summary":"Update acquisition/device parameter","description":"Update acquisition/device parameter.\n\nParameters\n----------\ndevice_id\n    Id of the device to be updated\npayload\n    Parameter dictionary\n\nReturns\n-------\n    Parameter dictionary\n\nRaises\n------\nHTTPException\n    404: Not found","operationId":"update_device_parameter","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"device_id","in":"path","required":true,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"string"}],"title":"Device Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/device/trigger_acquisition/{task_id}":{"post":{"tags":["devices"],"summary":"Trigger Acquisition","description":"Trigger an MRI acquisition for the given task.\n\nFetches the task from the exam manager, looks up the assigned sequence and device,\nsends the scan-start command via the device's open WebSocket, and marks the task as STARTED.","operationId":"trigger_acquisition","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"DeviceCreationRequest":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"status":{"anyOf":[{"$ref":"#/components/schemas/DeviceStatus"},{"type":"null"}]}},"type":"object","required":["name","description"],"title":"DeviceCreationRequest","description":"Device registration request pydantic model.\n\n(to be sent by user first adding the device to the platform)."},"DeviceOut":{"properties":{"device_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Device Name"},"serial_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Serial Number"},"manufacturer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Manufacturer"},"modality":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Modality"},"status":{"anyOf":[{"$ref":"#/components/schemas/DeviceStatus"},{"type":"null"}]},"site":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Site"},"parameter":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Parameter"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"id":{"type":"string","format":"uuid","title":"Id"},"datetime_created":{"type":"string","format":"date-time","title":"Datetime Created"},"datetime_updated":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Datetime Updated"}},"type":"object","required":["name","description","id","datetime_created"],"title":"DeviceOut","description":"Device pydantic output model."},"DeviceStatus":{"type":"string","enum":["ONLINE","OFFLINE","BUSY","ERROR"],"title":"DeviceStatus","description":"Pydantic definition of a commands."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"login"}}}}}}